New project: Circle-fractal drawing, in Javascript.
The project is written in HTML, JavaScript and CSS.

The project enables drawing a fractal based on circles using
web workers to avoid blocking or hindering the main event-loop.

It has not been tested on a variety of browsers, and may well
not work on several mainstream browsers due to usage of features
that may or may not be well supported.

It features:
- Drawing circle-fractal.
- Various settings and minor validation.
- Progress status in case drawing takes a long time (try iterations
  = 100).
- Cancelling drawing in progress and draw something different.